Set minimum expire date as minimum date for share expiration date input
Set minimum date of share expiration date input to the enforced minimum date instead of tomorrow.
Doesn't seem to have a real effect for now, probably because the min and max date parameters were previously
used for deactivating those dates in the calendar popup, which is no longer done in the latest stable nextcloud version.